// src/app/company-dto.ts
export class CompanyDTO {

    companyCode: string;
  cin: string;
  companyname: string;
  dateOfIncorporation: string;
  registerNo: string;
  telephoneNo: string;
  email: string;
  address: string;
  website: string;
  phoneNo: number;
  faxNo: string;
  city: string;
  state: string;
  pincode: number;
  logoName: any;
  createdBy?: string;
  updatedBy?: string;
  createdDate?: string;
  updatedDate?: string;
  active: boolean=true;
}